// tools/app/src/lib/api/action-types.ts
/**
    Cyberismo
    Copyright © Cyberismo Ltd and contributors 2024

    This program is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License version 3 as published by the Free Software Foundation.

    This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details.

    You should have received a copy of the GNU Affero General Public
    License along with this program.  If not, see <https://www.gnu.org/licenses/>.
*/

export type CardAction =
  | 'update'
  | 'delete'
  | 'create'
  | 'updateState'
  | 'createLink'
  | 'deleteLink'
  | 'editLink'
  | 'move';

export type AttachmentAction = 'add' | 'remove';

export type Action = CardAction | AttachmentAction;